+++ R/parameter_names.R
@@ -10,6 +10,8 @@
 
 parameter_names <- function(which) {
   
+  checkmate::assert_character(which, len = 1L)
+  
   which %>% 
     match.arg(c("constants", "enviro", "leaf")) %>%
     switch(
@@ -17,6 +19,30 @@ parameter_names <- function(which) {
          enviro = c("P", "r", "RH", "S_sw", "T_air", "T_sky", "wind"),
          constants = c("c_p","D_h0", "D_m0", "D_w0", "epsilon", "eT", "G",
                        "nu_constant", "R", "R_air", "s", "sh_constant")
+    )
+  
+}
+
+#' Names of parameters that can be provided as functions
+#'
+#' @inheritParams parameter_names
+#' 
+#' #' @examples 
+#' 
+#' tealeaves:::.parameter_functions("leaf")
+#' 
+#' @noRd
+
+.parameter_functions <- function(which) {
+  
+  checkmate::assert_character(which, len = 1L)
+
+  which %>% 
+    match.arg(c("constants", "enviro", "leaf")) %>%
+    switch(
+      leaf = NULL,
+      enviro = c("T_sky"),
+      constants = c("nu_constant", "sh_constant")
     )
   
 }