+++ tests/testthat/test-parameter-functions.R
@@ -3,30 +3,18 @@ library(tealeaves)
 
 test_that("some parameters can be defined as functions", {
 
-  .x <- list(
-    c_p = 1,
-    D_h0 = 1,
-    D_m0 = 1,
-    D_w0 = 1,
-    epsilon = 1,
-    eT = 1,
-    G = 1,
-    R = 1,
-    R_air = 1,
-    s = 1,
-    nu_constant = 1,
-    sh_constant = 1
-  )
+  error_message <- 'no applicable method for \'set_units\' applied to an object of class "function"'
+  blank_function <- function() {}
+  
+  # Constants ----
+  nms <- parameter_names("constants")
+  .x <- as.vector(setNames(rep(1, length(nms)), nms), mode = "list")
 
   cs <- constants(.x)
   expect_s3_class(cs, c("constants", "list"))
-  expect_length(cs, length(parameter_names("constants")))
 
   parameter_function <- c("nu_constant", "sh_constant")
-  parameter_numeric <- setdiff(parameter_names("constants"), parameter_function)
-
-  error_message <- 'no applicable method for \'set_units\' applied to an object of class "function"'
-  blank_function <- function() {}
+  parameter_numeric <- setdiff(nms, parameter_function)
 
   .x1 <- .x
   .x1[[sample(parameter_numeric, 1)]] <- blank_function
@@ -37,4 +25,42 @@ test_that("some parameters can be defined as functions", {
   for (i in parameter_function) .x1[[i]] <- blank_function
   expect_error(constants(.x1), regexp = NA)
 
+  # Environmental parameters ----
+  nms <- parameter_names("enviro")
+  .x <- as.vector(setNames(rep(1, length(nms)), nms), mode = "list")
+  
+  ep <- enviro_par(.x)
+  expect_s3_class(ep, c("enviro_par", "list"))
+  
+  parameter_function <- c("T_sky")
+  parameter_numeric <- setdiff(nms, parameter_function)
+  
+  .x1 <- .x
+  .x1[[sample(parameter_numeric, 1)]] <- blank_function
+  
+  expect_error(enviro_par(.x1), regexp = error_message)
+  
+  .x1 <- .x
+  for (i in parameter_function) .x1[[i]] <- blank_function
+  expect_error(enviro_par(.x1), regexp = NA)
+
+  # Leaf parameters ----
+  nms <- parameter_names("leaf")
+  .x <- as.vector(setNames(rep(1, length(nms)), nms), mode = "list")
+  
+  lp <- leaf_par(.x)
+  expect_s3_class(lp, c("leaf_par", "list"))
+  
+  parameter_function <- NULL
+  parameter_numeric <- setdiff(nms, parameter_function)
+  
+  .x1 <- .x
+  .x1[[sample(parameter_numeric, 1)]] <- blank_function
+  
+  expect_error(leaf_par(.x1), regexp = error_message)
+  
+  .x1 <- .x
+  for (i in parameter_function) .x1[[i]] <- blank_function
+  expect_error(leaf_par(.x1), regexp = NA)
+  
 })